Removes app settings for console verbosity
